perf(hireprocess): update hire process whole things again and write rounds page perfect
